requirements.txt
pandas==1.3.1
scipy==1.6.2
streamlit>=1.23.0
plotly.express
#altair

